backend/employeemng/requirements.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
gunicorn==23.0.0

